% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{quandl_api_key}
\alias{quandl_api_key}
\title{Get and Set an API key for Quandl}
\usage{
quandl_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{<character(1)> Quandl-provided API key for a registered account.}
}
\value{
If no input given, the currently set key. Note this will be NULL if
no key has been set yet. When setting a key, the key will be returned
invisibly.
}
\description{
Setting a key is necessary for almost all calls to Quandl. All registered
accounts get a key, even those not subscribing to paid bundles. Find your key
in your \href{https://www.quandl.com/account/api}{account settings}.
}
\examples{
\dontrun{
quandl_api_key("this-ismykey")
}
}
\references{
\href{https://docs.quandl.com/docs#section-authentication}{API documentation}
}
