% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{quandl_datatable_meta}
\alias{quandl_datatable_meta}
\title{Fetch metadata for a datatable}
\usage{
quandl_datatable_meta(code)
}
\arguments{
\item{code}{<\code{character(1)}> datatable code on Quandl}
}
\value{
<\code{list}> a list of names values of various types.
}
\description{
Quandl stores metadata on each datatable that includes names, descriptions,
when it was last updated, which columns can be filtered on, etc. Different
tables may return different fields.
}
\examples{
quandl_datatable_meta("WIKI/PRICES")
}
