% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{quandl_key_set}
\alias{quandl_key_set}
\title{Set your Quandl API key}
\usage{
quandl_key_set(api_key = NULL)
}
\arguments{
\item{api_key}{<character(1)> Quandl-provided API key for a registered account.}
}
\value{
<\code{character(1)}> the key, invisibly.
}
\description{
If no key is provided, will look in the environment variable
"QUANDL_API_KEY". If none found, will error.
}
\details{
Setting a key is necessary for almost all calls to Quandl. All registered
accounts get a key, even those not subscribing to paid bundles. Find your key
in your \href{https://www.quandl.com/account/api}{account settings}.
}
\examples{
quandl_key_set("this-ismykey")

# if you set your key in the "QUANDL_API_KEY" environment variable, that will
# be used automatically
quandl_key_set()
}
\references{
\href{https://docs.quandl.com/docs#section-authentication}{API documentation}
}
