% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyquandl-package.R
\docType{package}
\name{tidyquandl-package}
\alias{tidyquandl}
\alias{tidyquandl-package}
\title{tidyquandl: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClaytonJY/tidyquandl}
  \item Report bugs at \url{https://github.com/ClaytonJY/tidyquandl/issues}
}

}
\author{
\strong{Maintainer}: Clayton Yochum \email{claytonjy@gmail.com}

}
\keyword{internal}
