% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{quandl_datatable}
\alias{quandl_datatable}
\title{Retrieve data from a Quandl Tables API Endpoint}
\usage{
quandl_datatable(code, ..., .batch = 50L)
}
\arguments{
\item{code}{<\code{character(1)}> datatable code on Quandl}

\item{...}{filters and options to pass as parameters}

\item{.batch}{<\code{integer(1)}> maximum length of any parameter in a single
request; see Batching below}
}
\value{
<\code{tbl_df}> Results from Quandl in tibble form.
}
\description{
This is a replacementfor \link[Quandl:Quandl.datatable]{Quandl::Quandl.datatable} which allows for multiple
attempts, batches long parameters into multiple requests, always fetches all
results, and always returns a tibble.
}
\details{
Results are requested in CSV form and converted to a tibble via \code{\link[readr:read_csv]{readr::read_csv()}}.
}
\section{Batching}{

The Quandl API can only support a limited number of parameters in one call.
If we want to filter on e.g. 1,000 tickers, that requires multiple requests
to complete. This wrapper will handle this for you, automatically making
multiple requests to fetch the desired output. Currently, only one \code{...}
input can be longer than \code{batch_size}, so you can't filter on e.g. both 1000
tickers and 1000 dates.
}

\examples{
quandl_api_key(Sys.getenv("QUANDL_API_KEY"))

# get one day of prices on Apple from Wiki Prices
quandl_datatable("WIKI/PRICES", ticker = "AAPL", date = "2018-01-02")

# get one month of prices from two tickers
quandl_datatable(
  "WIKI/PRICES",
  ticker = c("AAPL", "MSFT"),
  date.gte = "2018-01-01",
  date.lt = "2018-02-01"
)

# only return some columns
quandl_datatable(
  "WIKI/PRICES",
  ticker = "AAPL",
  date = "2018-01-02",
  qopts.columns = c("date", "ticker", "close")
)
}
