% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{quandl_datatable}
\alias{quandl_datatable}
\title{Retrieve data from a Quandl Tables API Endpoint}
\usage{
quandl_datatable(code, ..., max_attempts = 3L, delay = 1L)
}
\arguments{
\item{code}{<\code{character(1)}> Datatable code on Quandl, specified as a string;
see \link[Quandl:Quandl.datatable]{Quandl::Quandl.datatable}}

\item{...}{Additional arguments passed to \link[Quandl:Quandl.datatable]{Quandl::Quandl.datatable}}

\item{max_attempts}{<\code{integer(1)}> Maximum number of times to attempt query.}

\item{delay}{<\code{integer(1)}> Number of seconds to wait between retries.}
}
\value{
<\code{tbl_df}> Results from Quandl in tibble form.
}
\description{
This is a wrapper around \link[Quandl:Quandl.datatable]{Quandl::Quandl.datatable} which allows for multiple
attempts, always fetches all results (\code{paginate = TRUE}), and converts the
result to a tibble.
}
\details{
Will throw error if empty table returned.
}
\examples{
Quandl::Quandl.api_key(Sys.getenv("QUANDL_API_KEY"))

# get metadata on Apple from Zack's Master Table
quandl_datatable("ZACKS/MT", ticker = "AAPL")

# get stock splits from Apple
quandl_datatable("ZACKS/HDM", m_ticker = "AAPL", action_type = 6)
}
