% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyquandl-package.R
\docType{package}
\name{tidyquandl-package}
\alias{tidyquandl}
\alias{tidyquandl-package}
\title{tidyquandl: A tidy interface to financial data via Quandl.com}
\description{
A tidy alternative to the `Quandl` package. Automatically batches and retries requests and ensures tibble outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ClaytonJY/tidyquandl}
  \item Report bugs at \url{https://github.com/ClaytonJY/tidyquandl/issues}
}

}
\author{
\strong{Maintainer}: Clayton Yochum \email{claytonjy@gmail.com}

}
\keyword{internal}
